+++ src/main/java/act/app/RuntimeDirs.java
@@ -21,10 +21,14 @@ package act.app;
  */
 
 import act.Act;
+import act.app.util.NamedPort;
+import org.osgl.util.S;
 
 import java.io.File;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 import static act.app.ProjectLayout.Utils.file;
 import static act.route.RouteTableRouterBuilder.ROUTES_FILE;
@@ -58,14 +62,25 @@ public enum RuntimeDirs {
         return file.exists() ? file : confBase;
     }
 
-    public static List<File> routes(App app) {
-        List<File> routes = new ArrayList<>();
+    public static Map<String, List<File>> routes(App app) {
+        Map<String, List<File>> map = new HashMap();
         File classes = classes(app);
-        routes.add(file(classes, ROUTES_FILE));
+        map.put(NamedPort.DEFAULT, routes(classes, ROUTES_FILE));
+        for (NamedPort np : app.config().namedPorts()) {
+            String npName = np.name();
+            String routesConfName = S.concat("routes.", npName, ".conf");
+            map.put(npName, routes(classes, routesConfName));
+        }
+        return map;
+    }
+
+    private static List<File> routes(File classes, String name) {
+        List<File> routes = new ArrayList<>();
+        routes.add(file(classes, name));
         File confRoot = file(classes, CONF);
-        routes.add(file(confRoot, ROUTES_FILE));
+        routes.add(file(confRoot, name));
         File profileRooot = file(confRoot, Act.profile());
-        routes.add(file(profileRooot, ROUTES_FILE));
+        routes.add(file(profileRooot, name));
         return routes;
     }